# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Expenses API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/expenses": {
            "post": {
                "description": "Register a expense into the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Create an expense",
                "parameters": [
                    {
                        "description": "Expense to be created",
                        "name": "expense_to_create",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/expenses.CreateExpenseInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Expense has been created",
                        "schema": {
                            "$ref": "#/definitions/entities.Expense"
                        }
                    },
                    "400": {
                        "description": "When a request does not fulfill need data",
                        "schema": {
                            "$ref": "#/definitions/validator.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Something unidentified has occurred"
                    }
                }
            }
        },
        "/expenses/to_paid": {
            "post": {
                "description": "Change a expense is_paid status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Set expense to paid",
                "parameters": [
                    {
                        "description": "ID to change to is paid",
                        "name": "expense_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/expenses.SetExpenseToPaidInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "When a request does not fulfill need data",
                        "schema": {
                            "$ref": "#/definitions/validator.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/expenses/toggle_is_paid": {
            "post": {
                "description": "Toggle is paid status from an expense",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Toggle is paid status from an expense",
                "parameters": [
                    {
                        "description": "ID to toggle is paid status",
                        "name": "expense_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/expenses.ToggleExpenseIsPaidInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/expenses.ToggleExpenseIsPaidOutput"
                        }
                    },
                    "400": {
                        "description": "When a request does not fulfill need data",
                        "schema": {
                            "$ref": "#/definitions/validator.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/expenses/update": {
            "put": {
                "description": "Updates an expense data depending if it is recurrent or not",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Updates an expense",
                "parameters": [
                    {
                        "description": "ID to toggle is paid status",
                        "name": "expense_update_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/expenses.UpdateExpenseInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense was updated"
                    },
                    "400": {
                        "description": "When a request does not fulfill need data",
                        "schema": {
                            "$ref": "#/definitions/validator.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/health_check": {
            "get": {
                "description": "It does a ping to db and returns service version",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "healt check"
                ],
                "summary": "Checks service health",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/controllers.HealthCheckOutput"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "$ref": "#/definitions/controllers.HealthCheckOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/login/google": {
            "post": {
                "description": "Creates a new session by token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login google"
                ],
                "summary": "Creates a new session by token",
                "parameters": [
                    {
                        "description": "Token generated by google oauth",
                        "name": "google_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Access information",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginOutput"
                        }
                    },
                    "500": {
                        "description": "Something unidentified has occurred"
                    }
                }
            }
        },
        "/recurrent_expenses": {
            "post": {
                "description": "Register into db a recurrent expense",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurrent expenses"
                ],
                "summary": "Create a recurrent expense",
                "parameters": [
                    {
                        "description": "Recurrent Expense",
                        "name": "recurrent_expense",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recurrentexpenses.CreateRecurrentExpenseInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "When a request does not fulfill need data",
                        "schema": {
                            "$ref": "#/definitions/validator.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/recurrent_expenses/all": {
            "get": {
                "description": "Finds all recurrent expenses registered in db",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurrent expenses"
                ],
                "summary": "Get all recurrent expense",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/recurrent_expenses/monthly_expenses": {
            "post": {
                "description": "Register expenses from all recurrent expenses for current month",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurrent expenses"
                ],
                "summary": "Create expenses from all recurrent expenses into db",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/reports/current_month": {
            "get": {
                "description": "Generates current month general details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Get current month details",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.LoginOutput": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "access_token_expires_at": {
                    "type": "string"
                },
                "user": {
                    "description": "RefreshToken          string    `json:\"refresh_token,omitempty\"`\nRefreshTokenExpiresAt time.Time `json:\"refresh_token_expires_at,omitempty\"`",
                    "$ref": "#/definitions/auth.UserData"
                }
            }
        },
        "auth.UserData": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "controllers.HealthCheckOutput": {
            "type": "object",
            "properties": {
                "db_status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "entities.Expense": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "day": {
                    "type": "integer"
                },
                "description": {
                    "$ref": "#/definitions/null.String"
                },
                "id": {
                    "type": "string"
                },
                "is_paid": {
                    "type": "boolean"
                },
                "month": {
                    "type": "integer"
                },
                "name": {
                    "$ref": "#/definitions/null.String"
                },
                "recurrent_expense": {
                    "$ref": "#/definitions/entities.RecurrentExpense"
                },
                "recurrent_expense_id": {
                    "$ref": "#/definitions/uuid.NullUUID"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/entities.User"
                },
                "user_id": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "entities.RecurrentExpense": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "$ref": "#/definitions/null.String"
                },
                "expenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Expense"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/entities.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entities.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "$ref": "#/definitions/null.String"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastname": {
                    "$ref": "#/definitions/null.String"
                },
                "name": {
                    "$ref": "#/definitions/null.String"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "expenses.CreateExpenseInput": {
            "type": "object",
            "required": [
                "amount",
                "name"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "for_next_month": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "expenses.SetExpenseToPaidInput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "expenses.ToggleExpenseIsPaidInput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "expenses.ToggleExpenseIsPaidOutput": {
            "type": "object",
            "properties": {
                "current_is_paid_status": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "expenses.UpdateExpenseInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
        "recurrentexpenses.CreateRecurrentExpenseInput": {
            "type": "object",
            "required": [
                "amount",
                "name"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "uuid.NullUUID": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if UUID is not NULL",
                    "type": "boolean"
                }
            }
        },
        "validator.ValidationError": {
            "type": "object",
            "properties": {
                "errors": {}
            }
        }
    }
}